migrate PAI and also move to shared where possible
