--- Content.Client/PAI/PAISystem.cs
@@ -1,8 +1,0 @@
-using Content.Shared.PAI;
-
-namespace Content.Client.PAI
-{
-    public sealed class PAISystem : SharedPAISystem
-    {
-    }
-}
