+++ Content.Server/PAI/PAISystem.cs
@@ -7,16 +7,13 @@ using Content.Shared.Interaction.Events;
 using Content.Shared.Mind.Components;
 using Content.Shared.PAI;
 using Content.Shared.Popups;
-using Content.Shared.Store;
-using Content.Shared.Store.Components;
 using Content.Shared.Instruments;
 using Robust.Shared.Random;
-using Robust.Shared.Prototypes;
 using System.Text;
 
 namespace Content.Server.PAI;
 
-public sealed class PAISystem : SharedPAISystem
+public sealed class PAISystem : EntitySystem
 {
     [Dependency] private readonly InstrumentSystem _instrumentSystem = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
@@ -38,8 +35,6 @@ public sealed class PAISystem : SharedPAISystem
         SubscribeLocalEvent<PAIComponent, MindAddedMessage>(OnMindAdded);
         SubscribeLocalEvent<PAIComponent, MindRemovedMessage>(OnMindRemoved);
         SubscribeLocalEvent<PAIComponent, BeingMicrowavedEvent>(OnMicrowaved);
-
-        SubscribeLocalEvent<PAIComponent, PAIShopActionEvent>(OnShop);
     }
 
     private void OnUseInHand(EntityUid uid, PAIComponent component, UseInHandEvent args)
@@ -106,15 +101,6 @@ public sealed class PAISystem : SharedPAISystem
         var val = Loc.GetString("pai-system-pai-name-raw", ("name", name.ToString()));
         _metaData.SetEntityName(uid, val);
     }
-
-    private void OnShop(Entity<PAIComponent> ent, ref PAIShopActionEvent args)
-    {
-        if (!TryComp<StoreComponent>(ent, out var store))
-            return;
-
-        _store.ToggleUi(args.Performer, ent, store);
-    }
-
     public void PAITurningOff(EntityUid uid)
     {
         //  Close the instrument interface if it was open
