+++ Content.Shared/Store/Events/IntrinisicStoreActionEvent.cs
@@ -2,6 +2,6 @@ using Content.Shared.Actions;
 
 namespace Content.Shared.Store.Events;
 
-public sealed partial class IntrinisicStoreActionEvent : InstantActionEvent
+public sealed partial class IntrinsicStoreActionEvent : InstantActionEvent
 {
 }
